*** empty log message ***

+++ src/bmz.h
@@ -9,7 +9,7 @@ typedef struct __bmz_config_data_t bmz_config_data_t;
 bmz_config_data_t *bmz_config_new();
 void bmz_config_set_hashfuncs(cmph_config_t *mph, CMPH_HASH *hashfuncs);
 void bmz_config_destroy(cmph_config_t *mph);
-cmph_t *bmz_new(cmph_config_t *mph, float c);
+cmph_t *bmz_new(cmph_config_t *mph, double c);
 
 void bmz_load(FILE *f, cmph_t *mphf);
 int bmz_dump(cmph_t *mphf, FILE *f);
