+++ src/bmz8.h
@@ -9,7 +9,7 @@ typedef struct __bmz8_config_data_t bmz8_config_data_t;
 bmz8_config_data_t *bmz8_config_new();
 void bmz8_config_set_hashfuncs(cmph_config_t *mph, CMPH_HASH *hashfuncs);
 void bmz8_config_destroy(cmph_config_t *mph);
-cmph_t *bmz8_new(cmph_config_t *mph, float c);
+cmph_t *bmz8_new(cmph_config_t *mph, double c);
 
 void bmz8_load(FILE *f, cmph_t *mphf);
 int bmz8_dump(cmph_t *mphf, FILE *f);
